orchestration/pyproject.toml
[project]
name = "lakehouse-cli"
version = "0.1.0"
description = "Enterprise CLI to run Lakehouse jobs on GCP (Dataproc Serverless, GCS, etc.)"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.3",
  "pyyaml>=6.0.1",
  "pydantic>=2.7.4",
  "rich>=13.7.1",
]

[project.scripts]
lakehouse = "lakehouse_cli.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]